+++ EmployeesSalaries/EmployeesSalaries/Services/EmployeeSalaryService.cs
@@ -5,7 +5,7 @@ namespace EmployeesSalaries.Services
 {
     public class EmployeeSalaryService : IEmployeeSalaryService
     {
-        public double CalcSalary(IEmployee employee)
+        public decimal CalcSalary(IEmployee employee)
         {
             return employee.SalaryCalc.GetNetSalary();
         }
